requirements.txt
paho-mqtt
streamz
pydantic
datamodel-code-generator
environs
pandas
pandasgui
networkx
environs